refactor: 재시작 기능 추가로 인한 handler 분리
